fix: implement deferred review items
- vfs: honor O_APPEND/O_TRUNC and status_flags at open/write; real
  truncate for tmpfs/fat32 (tmpfsTruncate, fat32.truncateFile); reject
  fake truncate on read-only files; require writable descriptors
- signals: sendSignal wakes blocked tasks (unblock+kick); semop/
  timerfd/epoll/futex/posix_mq wait loops check pending signals after
  resume (fatal -> exitTask, else -EINTR); signal_mask widened to u64;
  sigaction reads 8-byte sa_mask when the buffer is large enough;
  sigprocmask accepts sigsetsize=8; atomic rtSigtimedwait
- ipc: pipe EPIPE/SIGPIPE when the read end is closed (per-end
  refcounts); timerfd re-block loop and missed-expiration accumulation;
  sysv_sem SETVAL wakes waiters; posix_mq mq_close + one-shot mqNotify
  + deadline-bitmap timeout driven from the scheduler tick
- net: AF_UNIX wired into connect/sendto/read/write; DNS/DHCP register
  ports and pump RX while waiting; broadcast send uses the broadcast
  MAC; netif returns the DHCP address when configured; fd limits use
  vfs.MAX_FDS; UDP per-port refcount across fork; tcpShutdown clears
  listen slots; sendto EDESTADDRREQ; accept/recvmmsg validation; epoll
  reports real udp/eventfd state
- fs: execve honors user argv[0] (Linux semantics, sh.c updated); AIO
  rejects negative offsets and credits delivered bytes; file_io returns
  first-chunk errors; fat32 setFATEntry propagates read failures;
  readahead double-insert guard; wb_tick under lock; ext2 refreshes
  stale slot sizes on read
- drivers: NVMe admin queue lock
- boot/panic: panic parks APs via flag checked in the timer tick and
  falls back to lock-free serial writes on recursion; drop dead x86
  symbol table (344KB BSS); mapAcpiRegion becomes a checked no-op
- sync: fix unused RwLock/SeqLock/TicketSpinlock defects
- cleanup: delete sk5.zig, boot_info.zig, empty tools/* dirs

+++ kernel/proc/task.zig
@@ -115,7 +115,9 @@ pub const Task = struct {
 
     /// Signal mask — blocked signals (bit N = signal N+1 is blocked).
     /// SIGKILL (9) and SIGSTOP (19) cannot be blocked.
-    signal_mask: u32,
+    /// 64-bit to match the rt_* sigset_t ABI; only bits 0-30 are meaningful
+    /// (pending_signals stays u32 — signals 1-31 are supported).
+    signal_mask: u64,
 
     /// Signal handler addresses. 0 = default (terminate for now).
     /// Index 0 = signal 1 (SIGHUP), ..., index 30 = signal 31 (SIGUSR2).
